+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1 under development
 
 - New #16: Add `sendContact` method.
+- New #17: Add `getUserProfilePhotos` method and `UserProfilePhotos` type.
 
 ## 0.1.0 June 10, 2024